# kvs-watch: subscribe/unsubscribe to events asynchronously