Per discussion in #6690 and #6694.  It's a nice job throughput bump.  

Before
```
Running throughput.py 2048 jobs
throughput:     147.0 job/s (script: 142.6 job/s)
Total of 14 seconds
Running throughput.py 4096 jobs
throughput:     131.3 job/s (script: 129.9 job/s)
Total of 32 seconds
Running throughput.py 8192 jobs
throughput:     67.5 job/s (script:  11.6 job/s)
Total of 703 seconds
Running throughput.py 16384 jobs
throughput:     33.3 job/s (script:   6.7 job/s)
Total of 2456 seconds
```

After
```
Running throughput.py 2048 jobs
throughput:     196.4 job/s (script: 191.9 job/s)
Total of 11 seconds for everything
Running throughput.py 4096 jobs
throughput:     169.8 job/s (script: 168.0 job/s)
Total of 25 seconds for everything
Running throughput.py 8192 jobs
throughput:     173.1 job/s (script: 171.8 job/s)
Total of 48 seconds for everything
Running throughput.py 16384 jobs
throughput:     152.7 job/s (script: 152.1 job/s)
Total of 109 seconds for everything
```

Fixes #6694